The main-x11.c now looks for both 8x8.bmp and 8X8.bmp  (This is so the unzip -aL works.)
z-config.h used to have several font size (which corresponds to term size) definitions for x11.  However, only one of these was used.  This has been expanded to defines for all 7 terms + default.  The sizes are chosen to give a range of terms that will look good on many screen sizes.  (This is also to show that you can change them ;-) )

It would be nice if this could be done via a menu like in the windows version.  (As compared to compile time constants.)