Added "themed objects".  Monsters will drop objects more "appropriate" to what they are.  Objects generated on a certain depth will be native to depths closer to that depth on average.  Monsters deeper in the dungeon drop less stuff.  The stuff on average should be better.
This patch will be fairly buggy.  If you notice large amounts of junk (boots cloaks, spikes etc) being created far from level 1 - file a bug report.

I'd like to know if the types of objects matched with monsters is done right.

This patch will change balance... it is an attempt to get rid of alot of junk in the dungeon.