Cleaned up the LSL main.c file.  This is untested.  I cannot seem to get SVGALIB to work on this machine at the moment.  Feedback would be nice.
Cleaned up the use of maid-x11.c There now is a header file maid-x11.h which must be included to use it.

Added proper ascii support to the projected view.  It now works fairly well.  It is quite fast now.  Only the 'center on player' option is too slow to use.  It uses the 16x16.txt font (user editable) which needs to be beautified a little bit.