Made it possible to comditionally compile the wrapper C files.  The wrapper files are automatically generated by SWIG in the makefile.  Note that I've had to make a few hacks to the interface files in order to get this to work.
Eventually this may mean that the wrapper files may not have to be included in the cvs archive.  (Once all makefile support this feature.)

Note that this change does not turn python on.  It just makes it possible to compile the wrapper C files, and thus check the interface files for mistakes.

One question: It doesn't seem to like converting pointers... I've had to make a few minor changes like commenting out angband_term[], and replacing SWIGPTR_p_object_type with a string.  Hopefully I guessed right.  Note that this change adds a few #ifdef SCRIPT #endif pairs to the interface files.  These match the ifdef's in types.h