Added what I call [O] combat v3.  This is very much like v2, however there is a different critical hit formula, and the damage calculation is slightly simpler.
The deadliness calculation is now linear - and the percentage is shown instead of the old value.  (This is much more easy to understand than a value used indirectly to look up a table.)

The main difference from the old combat method is that the sides of the damage dice are changed, not the number of dice.  (This only really affects the statistics of the result.)  Weapons with only one damage dice will behave differently to weapons with many.