Removed non-maximize mode, and replaced it with something very much like normal maximize mode, however the stats are varied somewhat at birth.  This means that the game now uses an option with the best parts of both previous options.
There is variability at the start - with the possibility to get higher than 'normal' stats, like in non-maximize mode.
There is variability at the end, where the differing races and classes get different bonuses to the maximum possible stats like in maximise mode.

(This method is much better than the half-maximise mode proposed on rgra because that method seems to take the worst aspect of both earlier options, rather than the best aspects.)

Also, the bug with the warrior experience total in the birth screen is fixed.