Completely replaced the underlying wilderness data structure.  The new method scales much better, and is enormously 'cleaner'.  This allows further seperation of player and cave data.
This change allowed/forced lots of cleanup in other areas of the source.

The 'cave_empty_grid' macro no longer checks for the player - the existance
of a player will later have to be done via a function.

You can now edit the WILD_VIEW #define to change how much of the wilderness
is visible at one time - this is no longer quite so hard-coded.

The in_bounds() macros have been converted back to function pointers...  It
is now possible to have non-connected regions of the wilderness in_bounds,
and the regions in between, out of bounds.