Huge rewrite of the quest code.  At the moment it compiles - but it is completely untested.  The loading /saving code doesn't even bother to try and convert to the new format because it is only half finished.  Currently, only the old 'random' quests are supported - but they are done via huge hacks.
The wilderness /quest interface needs to be redesigned to get an
approximation of the old fixed-quests to work.  (Perhaps that will be done
within the next week or so.)

This commit probably breaks lots of stuff.

Be warned - we are now going into pre-alpha level of stability in the cvs
tree.  It will be a long time before things settle down again.  Welcome to
the 2.7.x branch, and keep a backup savefile...

Steven