Towns / quests in the wilderness now use regions instead of a massively hacky use of cave[][].
The dungeon also uses regions - allowing multiple levels to concurrently
exist, although lots of things need to be fixed to allow that to work.
Dungeon levels are smart enough to handle the deletion of objects/ monsters/
fields contained within themselves when the level is no longer needed.

The cave[][] array has been removed.

This change will probably be somewhat buggy...