Added a majorly hacked-up version of the borg to the game.
Warning: This probably will not work - it compiles, but that's about it for now.

The borg does not understand fields (so traps, stores and glyphs are a problem.)Also many "cheats" are commented out - this will cripple the borg in places.

The borg (barely) uses the new map interface -this will later mean lots of complex parsing code can be removed.