In the quest messages I removed the reference to the place where the quest was given.  It was cluttering the message.
I removed the distance requirement for the ruin quests.  I added a constraint so the player can have only 1 ruin quest for every dungeon.

I removed the distance requirement for the relic quests.  If there was only 1 dungeon in range all artifacts would end up there.  The player could then find The One Ring at the bottom of a 1 dungeon.

I removed the code that blocked the player from getting too many relic quest as it was not save persistant.  With 100 artifacts the chance that the next relic is one that has already been found is low.  Now you can get relics until you are heading for a dungeon with a bottom deeper than 100.  After killing The Serpent you can then continue to pick up all the relics.
Maybe this last one is not such a hot idea.  The alternative is to save the variable saying that this quest is blocked.