Enlarge the overhead map to use the left hand side fully.  It started at COL_MAP instead of 1.
If the cursor on the overhead map is on a (known) wilderness quest that quest is displayed in the upper line.
If the cursor is on a wilderness dungeon the upper line shows which (known) town is closest and if it is guarded or not.
These last two changes are a bit hackish because they draw from quest.c but the code is in maid-grf.c

Change the rewards for a bounty quest to account for how many monster you killed.
Change the relic quest so you don't have to id the artifact where you find it.
The bounty quest message -Still looking for them?- was diversified into -Still looking for he/she/it/them?-
The bounty quest message does not say you have killed 0 when you have to kill only one monster.