What a few days without the internet can do:
There are some procedures that call borg_power() to swap items.  Borg_power has as a side effect that it resets bp_ptr.  But these procs would sometimes leave bp_ptr inconsistent with the unswapped state.  These procedures now call borg_power at the end to make sure that bp_ptr is consistent with reality.  The existing inconsistency was causing an error with borg_restock and a wilderness loop.
Tried to fix another wilderness loop.  There was a funny calculation to determine the best shop in town.  But as the borg visits all the shops anyway I changed the best shop into the closest, unvisited shop.
Put in some more GOAL_NONEs
The previous entry trying to fix message handling for monks was incorrect as it missed al the uniques if they were hit by some fancy monk move.  I fixed it by first testing the message for a suffix.
I tried to fix the dungeon loop where the borg would hover close to a non-moving monster just around the corner.  I put in another line of sight check to improve on this.
There was a wilderness loop where the borg would reach a town but leave it right away for some unexplored wilderness.  This won't happen anymore because now the borg will head for a shop when it realizes that it reached the town.