Resync
Added a 'secure' option for file i/o.  This creates a child process which handles all secure file io and the game drops all permissions.  The two communicate over pipes, and the required information is passed back and forth.  The child does not trust the main game, and validates all dangerous input.  This should allow the gtk port to work in a multi-user setup.  Currently, this is controlled by the SECURE #define in config.h  This should, however be also controlled by the ./configure build system, and turned on whenever needed.